refactor: 파일 구조 변경 사항 적용
