+++ app/src/main/java/com/issueissyu/fe/data/remote/dto/request/auth/PhoneSendCodeRequest.kt
@@ -1,4 +1,4 @@
-package com.issueissyu.fe.data.remote.dto.request
+package com.issueissyu.fe.data.remote.dto.request.auth
 
 data class PhoneSendCodeRequest(
     val phone: String,
