+++ app/src/main/java/com/issueissyu/fe/data/remote/dto/request/auth/RefreshTokenRequest.kt
@@ -1,4 +1,4 @@
-package com.issueissyu.fe.data.remote.dto.request
+package com.issueissyu.fe.data.remote.dto.request.auth
 
 data class RefreshTokenRequest(
     val refreshToken: String
